requirements.txt
#torch==2.1.2+cu121
accelerate==0.28.0
braceexpand==0.1.7
click==8.1.7
deepspeed==0.9.5
diffusers
easyocr==1.7.1
filelock==3.9.0
Flask==3.0.2
hjson==3.1.0
huggingface-hub
hydra-core==1.3.2
imageio==2.34.0
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.7.5
matplotlib-inline==0.1.6
ninja==1.11.1.1
numpy==1.23.5
omegaconf==2.3.0
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.0.3
peft==0.4.0
pillow==10.2.0
pip==23.3.1
protobuf==4.21.2
pyrootutils==1.0.4
pytorch-lightning==2.2.1
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==1.43.0
setproctitle==1.3.3
setuptools==68.2.2
shapely==2.0.3
six==1.16.0
smmap==5.0.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
tokenizers==0.14.1
tqdm==4.66.2
transformers==4.34.0
triton==2.1.0
typing_extensions==4.8.0
wandb==0.16.4
Werkzeug==3.0.1
wheel==0.41.2
xformers==0.0.23.post1
zipp==3.18.1
hydra
pyrootutils
torchdata